Use `make install` target
Instead of manually and always resolving to monad-linux/{bin,lib}, we
use the CMake `install()` command to generate a `make install` target
which places powershell, libpshost, and libpsnative in
${CMAKE_INSTALL_PREFIX}/{bin,lib} respectively. CMAKE_INSTALL_PREFIX
defaults to `/usr/local` which we will want upon actual package
installation. During development, we will override this to be the path
to monad-linux in the superproject's Makefile.

Also move test output so monad-native Jenkins job can pick it up.