Fix another Start-Process test
This test was not viable as it didn't retain the process object,
instead relying on Get-Process with the process name, which could fail
to get the correct process (since no PID was available).

Since it is pointless to test Start-Process without -PassThru, and that
test was the exact same code as the fixed first test, it was removed.