Move monad to psl-monad on GitHub
- Removed more pre-processor conditionals
- Loads libpsl-native
- Pruned wmi (rewrote history)

The wmi folder was massive, and we did not need 99% of it, so a `git
filter-branch --tree-filter` was performed to remove all but the single
necessary subdirectory before pushing to GitHub.

The psl-monad repository can no longer be updated from the existing
GitSD sync. When the PSW teams have a Git repo for the monad codebase,
we will apply our patches to it and delete this fork.