Move Git installation instructions to contributing
The readme is getting quite large, so in an effort to shorten it, the
Git installation instructions (which have nothing to do with PowerShell,
but are helpful for new users), are now in `CONTRIBUTING.md`. Thus, if
someone is totally new to Git, they will be far more likely to review
our contributing guidelines, and so read our rules and best practices.

Cloning the source code was cleaned up so that initializing the
submodules is done in a separate step. This simplifies the first step,
making it more likely to succeed (and narrowing the possible failures).