Update content to `copyToOutput` and `publishOptions`
These must be duplicated as we require the same files for both `build`
and `publish`. This is the requirement of the new schema, as `content`
was split.

Note that while each key can take an object, if given just a `string[]`
as we do here, it defaults to `include`.

See the new schema: https://github.com/aspnet/Announcements/issues/175